Remove the v2 API
