+++ normandy/recipes/api/v3/shield_identicon.py
