--- normandy/studies/tests/api/v2/test_serializers.py
@@ -1,17 +1,0 @@
-import pytest
-
-from normandy.studies.api.v2.serializers import ExtensionSerializer
-from normandy.studies.tests import ExtensionFactory
-
-
-@pytest.mark.django_db()
-class TestExtensionSerializer:
-    def test_it_works(self, storage):
-        extension = ExtensionFactory()
-        serializer = ExtensionSerializer(extension)
-
-        assert serializer.data == {
-            "id": extension.id,
-            "name": extension.name,
-            "xpi": extension.xpi.url,
-        }
